customers/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
pluggy==1.5.0
PyMySQL==1.1.1
PyJWT==2.8.0
pytest==8.3.3
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
